semana05/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
Markdown==3.3.4
psycopg2==2.9.1
